+++ components/account.tsx
@@ -1,137 +1,68 @@
-"use client"
-
-import { UserType } from "@/types/user";
-import { use, useEffect, useState } from "react";
+import Image from "next/image";
 import Button from "./ui/button";
-import newUsername from "@/lib/usernameGenerator";
-import universities from '@/data/universities.json'; // Importing options from JSON file
-import { signOut } from "next-auth/react";
 
-function fetchUser(email: string) {
-    return fetch(`/api/data/user/${email}`, {
-        method: "GET",
-    }).then((res) => res.json());
-}
+import { signOutAccountAction, changeUsernameAccountAction, getUserFromDatabase, deleteAccountAction } from "@/app/actions";
+import { useEffect, useState } from "react";
+import { UserType } from "@/types/user";
+import { Modal } from "./modal";
 
-export default function Account({ currentUser }: { currentUser: string }) {
-    const initialUser: UserType = {
-        email: "",
-        username: "",
-        university: "",
-        createdAt: new Date(),
-        image: "",
-        name: "",
-    };
-    const [user, setUser] = useState(initialUser);
-    useEffect(() => {
-        fetchUser(currentUser).then((user) => {
-            setUser(user);
-        });
-        setUsername(user.username);
-        setUniversity(user.university != null ? user.university : "Unknown");
-    }, [currentUser, user.username, user.university]);
-    const [username, setUsername] = useState(user.username);
-    const [university, setUniversity] = useState(user.university != null ? user.university : "Unknown");
-
-    function handleUsernameChange(e: any) {
-        const newUsernameVar = newUsername(user.email!);
-        fetch(`api/data/user/`, {
-            method: "PUT",
-            headers: {
-                "Content-Type": "application/json",
-            },
-            body: JSON.stringify({
-                email: user.email,
-                username: newUsernameVar,
-                university: user.university,
-            }),
-        })
-            .then(response => {
-                if (response.status === 200) {
-                    setUsername(newUsernameVar);
-                } else {
-                    console.error("Username change failed");
-                }
-            })
-            .catch(error => {
-                console.error("Error:", error);
-            });
-    }
-
-    function handleUniversityChange(e: any) {
-        const newUniversity = e.target.value;
-        fetch(`api/data/user/`, {
-            method: "PUT",
-            headers: {
-                "Content-Type": "application/json",
-            },
-            body: JSON.stringify({
-                email: user.email,
-                username: user.username,
-                university: newUniversity,
-            }),
-        })
-            .then(response => {
-                if (response.status === 200) {
-                    setUniversity(newUniversity);
-                } else {
-                    console.error("University change failed");
-                }
-            })
-            .catch(error => {
-                console.error("Error:", error);
-            });
-    }
-
-
+export default function Account({ current_user }: { current_user: UserType }) {
+    const [user, setUser] = useState<UserType>(current_user);
+    const [buttonDisabled, setButtonDisabled] = useState<boolean>(false);
+    const [isOpen, setIsOpen] = useState<boolean>(false);
     return (
-        <div className="flex flex-1 flex-col items-center justify-start px-8">
-            <div className="flex flex-row w-full justify-between items-center h-36">
-                <h2 className="text-4xl font-bold">{user.name}&apos;s account</h2>
-                <img className="w-20 h-20 rounded-full border-4 border-white" src={user.image} alt={user.name} />
+        <div id="account-container" className="relative group transition-all">
+            <div id="account" className="flex flex-row items-center rounded-full bg-neutral-00 gap-2 p-2 shadow-[inset_0_0px_8px_rgba(0,0,0,0.3)]">
+                <p className="text-lg hidden lg:block">{user.email}</p>
+                <Image src={user.image} alt={user.name} width={50} height={50} className="rounded-full h-12 w-12" />
             </div>
-            <div className="flex flex-1 flex-col gap-4 text-2xl items-center justify-center">
-                <div className="flex flex-row gap-4 text-2xl items-center">
-                    <label htmlFor="email">Email: </label>
-                    <p id="email">{user.email}</p>
-                </div>
-                <div className="flex flex-row gap-4 text-2xl items-center">
-                    <form onSubmit={handleUsernameChange} className="flex flex-row items-center justify-center gap-4">
-                        <label htmlFor="username">Username: </label>
-                        <p id="username">{username}</p>
-                        <Button>Change</Button>
-                    </form>
-                </div>
-                <div className="flex flex-row gap-4 text-2xl items-center">
-                    <label htmlFor="university">University: </label>
-                    <select name="university" id="university" value={university} onChange={handleUniversityChange} className="text-2xl text-white bg-black">
-                        {universities.map((option, index) => (
-                            <option key={index} value={option}>{option}</option>
-                        ))}
-                    </select>
-                </div>
-                <div className="flex flex-row gap-4 text-2xl items-center">
-                    <label htmlFor="createdAt">With us since: </label>
-                    <p id="createdAt">{(new Date(user.createdAt)).getFullYear()}</p>
-                </div>
-                <div className="flex flex-row gap-4 text-2xl items-center">
-                    <form onSubmit={() => {
-                        const savedUser = user;
-                        signOut({ callbackUrl: "/" });
-                        fetch("/api/data/user", {
-                            method: "DELETE",
-                            headers: {
-                                "Content-Type": "application/json",
-                            },
-                            body: JSON.stringify({
-                                email: savedUser.email,
-                            }),
-                        }).then(() => true)
-                    }} className="flex flex-row items-center justify-center gap-4">
-                        <Button>Delete account</Button>
-                    </form>
+            <div id="account-dropdown" className="w-full hidden absolute right-0 mt-0 rounded-3xl bg-white shadow-lg p-4 group-hover:flex transition-all">
+                <div id="account-drop-inner" className="w-full flex flex-col gap-4">
+                    <div id="account-drop-top" className="flex flex-row items-center justify-between w-full">
+                        <h2 className="text-3xl font-bold">Account</h2>
+                        <div className="flex flex-row gap-4">
+                            <form action={signOutAccountAction}>
+                                <Button>Sign Out</Button>
+                            </form>
+                        </div>
+                    </div>
+                    <div id="account-drop-bottom" className="flex flex-col gap-4">
+                        <div id="account-drop-image" className="flex flex-row items-center justify-center gap-4">
+                            <div id="account-drop-bottom-username" className="w-full flex flex-col gap-1">
+                                <p className="text-xl font-bold">Username: </p>
+                                <div id="account-drop-bottom-username-control" className="w-full flex flex-row items-center justify-between">
+                                    <p className="text-lg">{user.username}</p>
+                                    <Button onClick={async (e) => {
+                                        if (buttonDisabled) return;
+                                        setButtonDisabled(true);
+                                        const updated_user = await changeUsernameAccountAction();
+                                        const db_user = await getUserFromDatabase();
+                                        setUser(db_user);
+                                        setButtonDisabled(false);
+                                    }}>Change</Button>
+                                </div>
+                            </div>
+                        </div>
+                        <div id="delete button" className="flex flex-row items-center justify-center gap-4">
+                            <Button onClick={async () => {
+                                setIsOpen(true);
+                            }}>Delete Account</Button>
+                            <Modal isOpen={isOpen} onClose={() => { setIsOpen(false) }}>
+                                <div id="delete-modal" className="flex flex-col gap-4">
+                                    <h2 className="text-3xl font-bold">Delete Account</h2>
+                                    <p>Are you sure you want to delete your account?</p>
+                                    <Button onClick={async () => {
+                                        if (buttonDisabled) return;
+                                        setButtonDisabled(true);
+                                        await deleteAccountAction(user.email);
+                                        setButtonDisabled(false);
+                                    }}>Confirm</Button>
+                                </div>
+                            </Modal>
+                        </div>
+                    </div>
                 </div>
             </div>
         </div>
-    )
+    );
 }
