+++ components/svgs.tsx
@@ -106,7 +106,13 @@ export function HamburgerSvg({ solid }: { solid: boolean }) {
     )
 }
 
-export function CrossSvg({ solid }: { solid: boolean }) { }
+export function CrossSvg({ solid }: { solid: boolean }) {
+    return (
+        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-6 h-6">
+            <path strokeLinecap="round" strokeLinejoin="round" d="M6 18 18 6M6 6l12 12" />
+        </svg>
+    )
+}
 
 export function InfoSvg({ solid }: { solid: boolean }) {
     if (!solid) {
