+++ components/ui/button.tsx
@@ -10,7 +10,7 @@ export default function Button({
     onClick?: ClickHandlerAsync | ClickHandler
 }) {
     return (
-        <button onClick={onClick} className="border bg-black text-neutral-200 transition-all duration-200 rounded-full py-2 px-4 text-sm">
+        <button type="submit" onClick={onClick} className="border bg-black text-neutral-200 transition-all duration-200 rounded-full py-2 px-4 text-sm">
             {children}
         </button>
     );
